Move base.py to _partial.py
