+++ dask_ml/_partial.py
@@ -1,11 +1,82 @@
 from __future__ import absolute_import, division, print_function
 
+import os
+from abc import ABCMeta
+
 import numpy as np
+import six
 from toolz import merge, partial
 
-from dask.base import tokenize
+import dask
 from dask.delayed import Delayed
-from dask import sharedict
+
+
+class _WritableDoc(ABCMeta):
+    """In py27, classes inheriting from `object` do not have
+    a multable __doc__.
+
+    We inherit from ABCMeta instead of type to avoid metaclass
+    conflicts, since some sklearn estimators (eventually) subclass
+    ABCMeta
+    """
+    # TODO: Py2: remove all this
+
+
+@six.add_metaclass(_WritableDoc)
+class _BigPartialFitMixin(object):
+    """ Wraps a partial_fit enabled estimator for use with Dask arrays """
+
+    _init_kwargs = []
+    _fit_kwargs = []
+
+    def __init__(self, **kwargs):
+        missing = set(self._init_kwargs) - set(kwargs)
+
+        if missing:
+            raise TypeError("{} requires the keyword arguments {}".format(
+                type(self), missing)
+            )
+        for kwarg in self._init_kwargs:
+            setattr(self, kwarg, kwargs.pop(kwarg))
+        super(_BigPartialFitMixin, self).__init__(**kwargs)
+
+    @classmethod
+    def _get_param_names(cls):
+        # Evil hack to make sure repr, get_params work
+        # We could also try rewriting __init__ once the class is created
+        bases = cls.mro()
+        # walk bases until you hit an sklearn class.
+        for base in bases:
+            if base.__module__.startswith("sklearn"):
+                break
+
+        # merge the inits
+        my_init = cls._init_kwargs
+        their_init = base._get_param_names()
+        return my_init + their_init
+
+    def fit(self, X, y=None):
+        fit_kwargs = {k: getattr(self, k) for k in self._fit_kwargs}
+        result = fit(self, X, y, **fit_kwargs)
+
+        # Copy the learned attributes over to self
+        # It should go without saying that this is *not* threadsafe
+        attrs = {k: v for k, v in vars(result).items() if k.endswith('_')}
+        for k, v in attrs.items():
+            setattr(self, k, v)
+        return self
+
+    def predict(self, X, dtype=None):
+        predict = super(_BigPartialFitMixin, self).predict
+        if dtype is None:
+            dtype = self._get_predict_dtype(X)
+        if isinstance(X, np.ndarray):
+            return predict(X)
+        return X.map_blocks(predict, dtype=dtype, drop_axis=1)
+
+    def _get_predict_dtype(self, X):
+        xx = np.zeros((1, X.shape[1]), dtype=X.dtype)
+        return super(_BigPartialFitMixin, self).predict(xx).dtype
 
 
 def _partial_fit(model, x, y, kwargs=None):
@@ -73,7 +144,7 @@ def fit(model, x, y, compute=True, **kwargs):
 
     nblocks = len(x.chunks[0])
 
-    name = 'fit-' + tokenize(model, x, y, kwargs)
+    name = 'fit-' + dask.base.tokenize(model, x, y, kwargs)
     dsk = {(name, -1): model}
     dsk.update({(name, i): (_partial_fit, (name, i - 1),
                                           (x.name, i, 0),
@@ -81,7 +152,7 @@ def fit(model, x, y, compute=True, **kwargs):
                     for i in range(nblocks)})
 
     value = Delayed((name, nblocks -1),
-                    sharedict.merge((name, dsk), x.dask, getattr(y, 'dask', {})))
+                    dask.sharedict.merge((name, dsk), x.dask, getattr(y, 'dask', {})))
 
     if compute:
         return value.compute()
@@ -110,3 +181,25 @@ def predict(model, x):
     xx = np.zeros((1, x.shape[1]), dtype=x.dtype)
     dt = model.predict(xx).dtype
     return x.map_blocks(func, chunks=(x.chunks[0], (1,)), dtype=dt).squeeze()
+
+
+def _copy_partial_doc(cls):
+    for base in cls.mro():
+        if base.__module__.startswith('sklearn'):
+            break
+    lines = base.__doc__.split(os.linesep)
+    header, rest = lines[0], lines[1:]
+
+    insert = """
+
+    This class wraps scikit-learn's {classname}. When a dask-array is passed
+    to our ``fit`` method, the array is passed block-wise to the scikit-learn
+    class' ``partial_fit`` method. This will allow you to fit the estimator
+    on larger-than memory datasets sequentially (block-wise), but without an
+    parallelism, or any ability to distribute across a cluster.""".format(
+        classname=base.__name__)
+
+    doc = '\n'.join([header + insert] + rest)
+
+    cls.__doc__ = doc
+    return cls
