+++ dask_ml/cluster/minibatch.py
@@ -3,7 +3,7 @@
 """
 from sklearn import cluster as _cluster
 
-from dask_ml.base import _BigPartialFitMixin, _copy_partial_doc
+from .._partial import _BigPartialFitMixin, _copy_partial_doc
 
 
 @_copy_partial_doc
