+++ dask_ml/linear_model/perceptron.py
@@ -1,6 +1,6 @@
 from sklearn.linear_model import Perceptron as _Perceptron
 
-from dask_ml.base import _BigPartialFitMixin, _copy_partial_doc
+from .._partial import _BigPartialFitMixin, _copy_partial_doc
 
 
 @_copy_partial_doc
