+++ dask_ml/linear_model/stochastic_gradient.py
@@ -1,6 +1,6 @@
 from sklearn.linear_model import stochastic_gradient as _sg
 
-from dask_ml.base import _BigPartialFitMixin, _copy_partial_doc
+from .._partial import _BigPartialFitMixin, _copy_partial_doc
 
 
 @_copy_partial_doc
